requirements.txt
langchain>=0.1.0
openai>=1.0.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
pytest>=8.0.0
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0
requests>=2.31.0
typing-extensions>=4.9.0
pydantic>=2.6.1
websockets>=11.0.3
aiohttp>=3.9.0
uvicorn>=0.27.0 